+++ frontend/src/constants/urls.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = 'https://us-central1-clubhub-dev-89ca0.cloudfunctions.net/default';
