+++ frontend/src/styling/StyledHome.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { colors } from '../constants/Colors';
+import { colors } from '../constants';
 
 export const Container = styled.div`
     display: flex;
